tasks/testsuite: add a test suite for the Tasks service.
I intend to add a simple in-memory implementation alongside the
PostgreSQL-backed one. Mostly for fun, but it may also be valuable for things
like playing around with a UI for the service -- the in-memory one can prove to
be more lightweight and allow for easier experimentation.

To help keep the in-memory implementation up to spec, we make sure it behaves
consistently with the PostgreSQL-backed implementation, by running them both
through the same test suite. If the test suite is comprehensive, this should
make sure there is no meaningful difference between implementations as long as
they both pass the tests.

For the existing implementation, there is also a large benefit: the test running
time is greatly reduced. This is because instead of starting a Docker container
running PostgreSQL for each test, there is only one container, and the database
is cleared out between tests instead. We lose the ability to run tests in
parallel, unfortunately, but I think the tradeoff is worth it.
