+++ tasks/service/BUILD.bazel
@@ -27,17 +27,13 @@ go_test(
     data = ["//tasks/postgres:schema.sql"],
     embed = [":service"],
     deps = [
+        "//postgres",
         "//postgres/postgrestest",
         "//tasks/tasks_go_proto",
-        "@com_github_google_go_cmp//cmp",
-        "@com_github_google_go_cmp//cmp/cmpopts",
+        "//tasks/testsuite",
+        "@com_github_stretchr_testify//suite",
         "@org_golang_google_grpc//:grpc",
-        "@org_golang_google_grpc//codes",
         "@org_golang_google_grpc//credentials/insecure",
-        "@org_golang_google_grpc//status",
         "@org_golang_google_grpc//test/bufconn",
-        "@org_golang_google_protobuf//proto",
-        "@org_golang_google_protobuf//testing/protocmp",
-        "@org_golang_google_protobuf//types/known/fieldmaskpb",
     ],
 )
