+++ tasks/testsuite/BUILD.bazel
@@ -1,0 +1,23 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "testsuite",
+    srcs = [
+        "client.go",
+        "doc.go",
+        "testsuite.go",
+    ],
+    importpath = "go.saser.se/tasks/testsuite",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//tasks/tasks_go_proto",
+        "@com_github_google_go_cmp//cmp",
+        "@com_github_google_go_cmp//cmp/cmpopts",
+        "@com_github_stretchr_testify//suite",
+        "@org_golang_google_grpc//codes",
+        "@org_golang_google_grpc//status",
+        "@org_golang_google_protobuf//proto",
+        "@org_golang_google_protobuf//testing/protocmp",
+        "@org_golang_google_protobuf//types/known/fieldmaskpb",
+    ],
+)
