+++ tasks/testsuite/doc.go
@@ -1,0 +1,3 @@
+// Package testsuite contains a suite of tests that any implementation of the
+// Tasks service should pass.
+package testsuite
